Consider SPELL_ATTR1_UNAFFECTED_BY_SCHOOL_IMMUNE when checking school immunity. Cleanup MAX_SPELL_EFFECTS for-loops, value can be 0-2 so dont need to assign counter to int or uint32 :>

+++ src/server/game/Spells/Auras/SpellAuraEffects.cpp
@@ -3892,7 +3892,7 @@ void AuraEffect::HandleAuraMounted(AuraApplication const * aurApp, uint8 mode, b
         CreatureModelInfo const *minfo = sObjectMgr->GetCreatureModelRandomGender(displayID);
 
         //some spell has one aura of mount and one of vehicle
-        for (uint32 i = 0; i < MAX_SPELL_EFFECTS; ++i)
+        for (uint8 i = 0; i < MAX_SPELL_EFFECTS; ++i)
             if (GetSpellProto()->Effect[i] == SPELL_EFFECT_SUMMON
                 && GetSpellProto()->EffectMiscValue[i] == GetMiscValue())
                 displayID = 0;
